[427799] If opening failed, make sure the cleanup is actually done
If an the call to open() is cancelled by an exception, we try to cleanup
by calling close(). However close() returns immediatly without doing
anything if the session is not opened, as returned by isOpen(), but it
is only marked as such right at the end of open(). If the exception
occurs before that, calling close() in the finally block is ineffective.

Bug: 427799
Change-Id: I0b919fae2c069189d6e018f2148b4bdeefd1d4f5
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>