[41455] Coerce non-null values to true/false when a Boolean is expected
If <%self.someRef%> has been commonly used in .odesign files for years
even when someRef has nothing to do with a boolean. To ease the
migration of .odesign files it is sensible to keep this type coercion
mechanism. Not having this behavior means that when migrating away from
<%%> the specifier can no longer rely on batch search/replaces as
aql:self.someRef will return false when someRef is not a boolean.

Bug: 41455
Change-Id: I66bb9cdd5a5a9c9dfa2cca54316ff4f5ac61e37d
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>