[461498] Systematically declare Dialect Metamodels as available
There was no guarantee at evaluation time that Ecore models used by
Sirius for representations or VSM models would be accessible in the
context of a Session. This is especially problematic for IInterpreters
implementations which will not do their own lookup in the EMF
EcorePackage registry when a type is not found (AQL notably won't do any
lookup and expect the client code to declare the EPackage to support).
This commit changes this by using the org.eclipse.sirius.mmdescriptor
extension point in each dialect to declare the EPackages used by the
host dialect, making those EPackages usable in a interpreted expression.

Bug: 461498
Change-Id: Ie90ff18040894d27b5714795ee162ad277816dc4
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>