[460706] Add a generic dialog to browse in workspace and plugins
Define a generic dialog to browse for files in the workspace and
the plugins. A list of extensions can be given as arguments to filter
the files. The plugins that are searched in are the plugins that are in
the target platform.

A helper to use this dialog for searching images is defined in the
dialog's class.

Add dependencies to org.eclipse.ui.ide and
org.eclipse.equinox.frameworkadmin in org.eclipse.sirius.editor.

Bug: 460706
Change-Id: I3fe194458ec50bfdc536e1ad13cd143f0339f175
Signed-off-by: Benjamin Grouhan <benjamin.grouhan@obeo.fr>