[456351] Modify test to keep it working
Before 456351 migration, we already had resources under resourceSet
corresponding to an unresolvable resource URI. That resource was created
when attempting to resove proxyURI of DAnalysis.getModels.

After 456351 migration, the resolution of resource URI is done directly
in getModels implementation.
So now just calling getModels is sufficient to add that unresolvable
resource to resourceSet. 
That's why the test
SiriusControlTest.testUncontrolWithTwoRepresentationsFileLinkedToTheUncontrolledElement
is KO.
But, as we continue to accept such resource, we just modify the test to
keep it working.

Bug: 456351
Change-Id: I5133407e3bd8c2aa9daf7d788eba3f7155215789
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>