[465429] Change the type of "self" after a CreateView
CreateView is defined in the Diagram dialect and not in the generic
part. Because of that this commit exposes a protected method in
AbstractInterpretedExpressionQuery so that
DiagramInterpretedExpressionQuery can also switch the type of "self"
during the analysis.

The type of "self" after a CreateView is set to the specific type
depending on the mapping type and values (DNode, DEdge, DNodeList or
DNodeContainer)

Bug: 465429
Change-Id: Ia19fe04c866d0027cda76907d84b467a69ef8322
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>