[461593] Decorate and manage container and list labels in outline and hide wizard
Update the outline label provider to change the font and decorate the
icon of list and container when their label is hidden. The node and edge
are not impacted as node with "node label" does not support hide label
yet and node with bordered label and edges have their labels displayed
as an element in the outline (the user can select and move them on the
diagram).

Update the content provider used in the show/hide wizard to display the
label of containers and lists as virtual nodes in the wizard. The choice
has been made to not add this virtual node in the outline as the user
cannot select or move those labels.

Bug: 461593
Change-Id: If521958576ad4a286cd93be771755fe37bf7c28d
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>