[464958] Make LogThroughActiveDialectEditorLogListener more robust
LogThroughActiveDialectEditorLogListener is installed by
SiriusEditPlugin.start() hence is pretty much always running.
Reworking the conditions when it reacts to a logged exception so that it
will only call PlatformUI.getWorkbench() for exceptions which are of
interests (and not call the workbench and then check if the exception is
of interest). Also check for the workbench running before accessing it
to avoid situations where : 
 java.lang.IllegalStateException: Workbench has not been created yet.
 
would be thrown.

Bug: 464958
Change-Id: I1dba32be04e503ae6995665af0a28b6575ff4ab2
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>