[466456] Ignore resized border node during relocate of brothers
During the conflict detection after a resize, the resized node should be
ignored for other border nodes.
But this reveals another problem (detected with tests of org.eclipse.sirius.tests.swtbot.uml.UmlPortDragAndDropTest). When computing the realExpandedBounds
(SpecificBorderItemSelectionEditPolicy.getRealExpandedBounds()), the
borderItemHasMoved state become wrong. A new API has been added in the
DBIL to reset this state in this case.

Bug: 466456
Change-Id: I94397d579df638b98466ca7b15c25a83d20a35d7
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>