[466455] Replace RefreshIDFactory static calls with a scoped holder
Now that MinimalEObjectImpl is being used for Sirius's model the
approach of attaching IDs to every EObject through eAdapters() became
more CPU intensive than desired representing 17 to 18% of the time spent
refreshing a diagram (You can check the logic in
org.eclipse.emf.ecore.impl.MinimalEObjectImpl.eAdapters()) 

This commit changes this approach by introducing a non-static holder for
such IDs which is scoped to a given representation (RefreshIdsHolder).
The holder itself still is an adapter and clear its state when its being
detached from the representation (on Resource.unload for instance).

This commit changes both the Diagram and Table algorithms which were
making use of the RefreshIDFactory.

Notable change of this (besides some API) is that a RefreshID is now
unique only in the scope of given representation - which can actually be
expected from its name.


Bug: 466455
Change-Id: I2aca1c435bb17de42e972ea8c3061eda18834460
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>