[perfs] Make EqualityHelper fails faster based on the fragment URIs
This commit adapts the EqualityHelper code so that we can detect both
objects are not the same (hence, will have a different uri) sooner
without building the whole URI.

Instead browse from the current EObjects to the containers and check the
URI fragments at each step. Once the browsing ends up at the Resource
level, only check for the resource URI then.

Change-Id: Ib9b2798ce1433160f1ea8910ef867da766f13398
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>