[perfs] Uses Sets for collection which are only asked for contains()
This can represent up to 10% of the time spent in a diagram refresh.

Also uses Iterables.transform() instead of Collections2.transform()
based on the Javadoc : "When a live view is not needed, it may be faster
to copy the transformed collection and use the copy." and move the
"transformed" Set into a conditional branch to avoid trigerring an
instantiation of Set.

Change-Id: I0d7eb981f9e00543c8eb49551ad5216b1a92f1f7
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>