[441946] Reorder the sub-menus of the contextual menu in the VSM.
Use priorities added in the plugin.properties file of the sirius.ui
plugin to reorder the sub-menus. Currently the values are all set to the
same arbitrary great number (10⁹), so the actual order of sub-menus
should not change.

Also add a helper class (EClassHelper) defining a static method to get
the path of an EClass object. The path returned by this method is used
as a key to retrieve the priority in the plugin.properties file.

Last, add a test checking that all EClass contained in the description
package of each metamodel's dialect have a priority associated in the
plugin.properties file.

Bug: 441946
Change-Id: Ibab27d13faf7f2d83cb249d771e5a4a06ac98fe9
Signed-off-by: Benjamin Grouhan <benjamin.grouhan@obeo.fr>