[436359] Specific label border styles for list and container
Two new label border styles are now available. 

"nolabelBorderForList" allows to define lists without line separator
between the title and the first list item.

"labelBorderForContainer" allows to define a container with the same
title area than a list. Its content pane is also impacted it will not be
possible to move children elements in this area.

Bug: 436359
Change-Id: Ibf1b2efd20ad2f009046debdd2814f8cd2ca4c8c
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>