[436833] Introduce a Sirius-Aware Properties Edition Component
This PropertiesEditionComponent will make sure not to install/uninstall
the adapter used by EEF to listen for changes but to reuse the same
instance.

A SessionManagerListener is introduced to uninstall the adapter when the
session is getting closed.

Note that this change uses APIs provided by EEF 1.5 and requires an
upgrade then.

Bug: 436833
Change-Id: Id08788abfbcc930fc8091f7ec38eefff19aa4672
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>