[467115] ModelExplorerView uses LockDecorationUpdater to be refreshed
This allows to automatically refresh this view when the
IAuthorityListeners are notified about lock status changes.

1- During the init of the ModelExplorerView, the LockDecorationUpdater
resgister this view. A SessionManagerListener is added to detect the add
or the remove of session (it is removed during the dispose of the view).
2- This PermissionAuthoritySessionManagerListener adds the
LockDecorationUpdater as PermissionAuthority on the new session (and
remove it during the remove of the session).
3- The LockDecorationUpdater launches a refresh of ModelExplorerView,
during lock/unlock notifications, in a job to avoid deadlock.

Bug: 467115
Change-Id: I20604d8ed500159fb640813b16b1613adb204e8f
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>