[436359] Prepare the region collapse support
Add the drawer style used by the compartiment collapse handle and the
resizable policies to control the compartiment collapse.

This commit is a preparation step, all diagram with regions created
after this commit will have a GMF model ready to support the collaps of
regions.

It also add a migration step to prepare previously created diagrams with
regions by adding the drawer style on the region compartiment nodes.

Bug: 436359
Change-Id: I31e900f15fb2ce03862730cdae531b85b8bd6a65
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>