[436359] Refactor the experimental queries names
The modified queries have been experimental for 2 years and are stables. 

In org.eclipse.sirius.diagram.ui, the internal DNodeContainerQuery keeps
its name to be consistent with the other queries in the same package, it
now inherits from the same named query from org.eclipse.sirius.diagram
to ease the use of both queries in the diagram ui code.

Bug: 436359
Change-Id: Iabbdf1f845f500c57f64aaec287cd7b075a74e6b
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>