[456351] Keep semantic resources URIs relative to aird
Semantic resource URI serialization has to be made relative to aird URI
when both resources are in same project.
As the semantic resource URI has to be absolute in workspace for proper
use, the URI is resolved/unresolved during
deserialization/serialization.
We keep the same behavior of serializing the relative path than for
DAnalysis.models before.

Bug: 456351
Change-Id: I0a0d0a9e06c30aa33cee5e537ca7d3c2a621bd04
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>