[441946] Refactor the way menu items are associated with priorities
For subtle reasons due to the lifecycle of the MenuManager and IActions,
the previous method only worked for the first menu invocation. Refactor
the code with less assumptions about which actual instances of IAction
are created/reused/re-created/etc.; use the more reliable type keys
computed by EClassHelper.getPath() to track priorities instead.

Bug: 441946
Change-Id: Ib9317a036e325ed0cd5b18f930f1fd6fe04a1795
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>