[441946] Improve the ordering of menu items
Define the actual order of VSM context menu categories and items inside
categories. Not all items may by fully handled, but it is already way
better than the previous state.

Also make the menu priorities test less strict: it should not be a
failure if a key is missing (the code is ready to handle this with
default priorities), but only if a key is malformed.

The "missing key" tests is still useful, but only as an optional report
which can be used to reviews if some important types have been
forgotten/left without priority.

Bug: 441946
Change-Id: I4aa8b521dfbae98fcfbf9b932ed292b8dc407f8c
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>