[468021] Fix pb on edges after a drag'n'drop of a node
Since, the bug 427017, when a node containing sub-nodes (border or not),
that have edges, is drag'n'dropped, these edges become unstable.
In commit [1], ViewUtil.destroy(view) has been replaced by
EcoreUtil.remove(view) to avoid the GMF crossReferencer instantiation.
But ViewUtil.destroy(view) also destroys edges without source or target
after the destroy of the view.
This is now done in AbstractCanonicalSynchronizer.deleteViews().
The case of NoteAttachment, with null element, is now handled in
DDiagramCanonicalSynchronizer.

[1]
http://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=94b51052cdab27eb399b7c7655c66fecb50a3648

bug: 468021
bug: 427017
Change-Id: Ib88c6fc7b85cbfccae172aebce01ba8594141af5
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>