[467508] Delete impossible after external modif of representations file
The commit [1] has a side effect. It adds the AirDCrossReferenceAdapter
to the ResourceSet instead of to the Resource. So during the reload
(unload) of the resource, in the issue scenario, the
AirDCrossReferenceAdapter is removed from the Resource. But when GMF is
called during delete through ViewUtil.destroy(View), it uses the
AirDCrossReferenceAdapter returned by the ResourceSet, but this last is
not correct as the Resource is not considered. Before this commit, with
the same scenario, GMF finds no CrossReferenceAdapter so it creates a
new one.
Now, the AirDCrossReferenceAdapter of the ResourceSet is added during
the reload (in ReloadRepresentationsFileCmd).

[1]
http://git.eclipse.org/c/sirius/org.eclipse.sirius.git/commit/?id=5b294ef32d426ab47c5339379f6a293a8d75044d

Bug: 467508
Change-Id: I0d690d9538e82231d0030a048a3e03e58cce38f7
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>