[461593] Container and list related figure adaptation for hidden labels
Since commit 77d6b28496cd23c13e37484c66b12358b9f33bf4, the compartment
border/margin and the label offset are set regarding the label
visibility state.

This commit fix a regression for Operand in Sequence diagrams in which
the OperandFigure is created with a 0 pixel one line margin border and
must stay with this 0 pixel margin to avoid changes in existing Sequence
diagrams.

Bug: 461593
Change-Id: I22fcf41c556d82362d106c8b10a9583292f35a00
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>