[475820] Disable snap on Sequence diagrams
Force false values for the GMF diagram workspace snap
preferences of Sequence diagrams. A property change listener is added to
asynchronously reset the snap properties to false when something tries
to set them to true: this allow to wait that all property change
listener have been notified of the first change before doing the second
one. Without this, the Rulers and Grid property section displays the
snap as enabled when the user tries to activate it even if we maintain
it disabled.

Overriding getAdapter in all SequenceLayout/Reorder sensitive edit parts
have also be tested but without  decisive effect: the snap command is
still executed and call a reorder during arrangeAll or other actions.

Bug: 475820
Change-Id: I08c6dacec505b8826912eca18e051769e69eca82
Cherry-picked-from: 475820
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>