[467115] Remove LockedModelExplorerTest from tests suite
After long analysis, this test really reveals a problem. It fails only
if it is launched after another test that opens the Sirius perspective.
That other test, by opening Sirius perspective, also opens the
ModelExplorer view in another perspective than Modeling one. So when the 
org.eclipse.ui.IWorkbenchPage.hideView(IViewReference) is called on
ModelExplorer view, the dispose method is not called and the refresh job
is launched (even with ModelExplorer view closed).

We can not use the dispose() method for bug 467115. It seems that
IPartListener2 is the solution with the
partHidden(IWorkbenchPartReference) and
partVisible(IWorkbenchPartReference). But the first tests are not
conclusive. More analysis is needed (will be done after Sirius 3.0).

The intial problem of bug 467115 is solved, because if the ModelExplorer
view is never opened (case of customer that customizes Sirius), the job
is never called.

This commit also removes the test API, added just for this test, that is
not really reliable.

Bug: 467115
Change-Id: I27a4a209cf8482b4b8ceb7535727e1e94e8bdc5b
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>