[467115] Fix LockedModelExplorerTest by closing the Sirius perspective
The Sirius perspective is now closed duting the setUp step. If it was
opened by another test before, the ModelExplorer view is opened in 2
perspectives and the close view made in this test has not the expected
result.
As explain in the previous commit for bug 467115, this reveals a real
problem (bug 469848), a minor one, that will be fixed in another commit.
        
Bug: 467115
Change-Id: I827eb6de2d09963bae40bf0c231a763e5004fb7d
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>