[470770] Improve corner dimension update for regions
Imrove rendering of the region corners when the region container has
rounded corner. The current commit is not a full correction: it just
improve the global rendering, but moves the issue to the rendering of
the first/last regions only: we should be able to choose the corner we
want to draw rounded.

Update the corner dimension of all GradientRoundedRectangle without
checking if the current style is a FlatContainerStyle. The choice to
support or not the rounded corner is now only done in
AbstractDiagramElementContainerEditPart.createNodeShape() where the
figure is created.

Bug: 470770
Change-Id: Id91cd8a88be319f2a0042757d00344167df26e4c
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>