[463485] Update SWTBot framework for edge feedback and shortcut
For tests of bug 463485, we need 2 additional points:
* to allow a key pressed during a drag'n'drop,
* to move bendpoint of edge.

For the first, allow a key pressed during a drag'n'drop, the method
mouseDragWithKey is added in SWTBotSiriusFigureCanvas. This method has
an additional parameter to precise the key code to press during the
drag'n'drop. It is possible to call this method with SWT.None to
retrieve the standard behavior of mouseDrag method.

For the second point, move bendpoint of edge, the behavior
is currently not correctly handled. Indeed, the feedback of edge is not
redrawn, so the policy/command uses wrong data when it creates/executes
the SetConnectionBendpointsCommand. The method
mouseDragWithKey has been completed to force an update of the
Viewport (that draws the feedback) and then to dispatch a second time
the mouse drag event.

Bug: 463485 
Change-Id: If2709b01057b1db88e85fddeca0ca8f725ad8f41
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>