[461590] Border line style customization MM update
Add a border line style attribute to the BorderStyleDescription and
BorderStyle EClasses.
Display the borderLineStyle in the Style property tab of the Sirius
diagram editor.
Hide the BorderedStyle.borderSizeComputationExpression from the Style
property tab of the Sirius diagram editor as changing it has no impact:
the style's expression is not evaluated whereas the description's
expression is.

Bug: 461590
Change-Id: Ie8bc0e324c640abb6a93ec22a24250b2448fe052
Signed-off-by: Belqassim Djafer <belqassim.djafer@obeo.fr>
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>