[465211] Allow container and list without border
* Remove code that replaces hardly "0" by "1" for border size of
container.
** Also change the outline aspect of figure that is problematic on
Windows OS. 
* Add a migration participant for borderSize and
borderSizeComputationExpression to avoid a change of behavior for
ContainerMapping with 0 as borderSize.
* Change default value for Gradient and Parallelogram style in VSM
editor. The default value is now 1 for these new kinds of style. 0 is
kept for new Workspace Image style as it is considered as normal to not
have a border on container represented with an image.

Bug: 465211
Change-Id: Iee17b4abd00e87214f01b6a775bc448f8d90e185
Signed-off-by: Belqassim Djafer <belqassim.djafer@obeo.fr>
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>