[472057] Externalize translatable strings from plugin.xml files
Also Configure PDE Tooling to consider non-externalized strings in
plugin.xml as errors from now on.

I've reviewed all our own extension point definitions (*.exsd) for
possibly translatable strings which were not marked as such (and thus
for which we would not get the benefit of PDE's check), but the only
candidates are for attributes we do not actually use anywhere (e.g. we
allow an extender to provide labels for
org.eclipse.sirius.common.acceleo.mtl.importhandler, but even if it is
supplied we do nothing with it). To avoid changes which would have no
actual benefit, the *.exsd files have not been modified.

Bug: 472057
Change-Id: I51f932c8ae5ee8bfea268213649fc3448d0bd827
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>