[453037] Do not restore session
Call getSession(false) instead of getSession() in places where the
session is closed and we want to avoid to recreate and open a new
session. Typically in CustomSiriusDocumentProvider: the session has been
closed, the ui session tries to close all opened editor but check if the
editor is dirty: this checks ask to retrieve the session.

Bug: 453037
Change-Id: Ic5a04975f7bea17b69a47390fd89e9c3fc48b166
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>