[463662] Label size minimal value set to 1
A label size of 0 pixel should be forbidden. Indeed, SWT does not handle
a Font with 0 pixel.

The label size minimal value of style in Diagram, Table and Tree
descriptions
have be changed from 0 to 1. The same change has been applied to user
label size customization.
Forthermore, if the label size is programatically set to 0, the
displayed font will still be 1.

Bug: 463662
Change-Id: I5d3f3217b818cc38e3764708a23a6241a85778df
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>