[465328] Change edge labels location according to edge move
* The concerned policies have been updated to store the initial points
(with InitialPointsManagerForEdgePolicy). These initial points are then
used to compute the new label location with EdgeLabelsComputationUtil by
the concerned Command.
* A new LabelLocator, EdgeLabelLocator, is added to have a correct label
feedback during the display of the edge label feedback. This
LabelLocator must be initialized with method setFeedbackData (and then
reset with method eraseFeedbackData) according to the feedback of the
edge lifecycle (Policy#show*Feedback and
Policy#eraseConnection*Feedback).

Bug: 465328
Change-Id: Icbe070a26936eb141189be0d5e1316a5a372abe1
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>