[435507] Fix test for edges with border nodes
The test class EdgeWithBorderNodeCreationPositionWithSnapToGridTest has
never been added to test suite for bug 435507. It reveals a bug that
is fixed with this commit.
The bug is that if during the edge creation, there is no intersection
between the snapped creation points and the source (or target) figure,
the result is not a snapped one.

Bug: 435507
Change-Id: I445e5fa05d41b861814dd0f3656a93faee14a59c
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>