[470770] Introduce RegionRoundedGradientRectangle
Since Bug 465211 and commit dbecde152cd2c577da12be3ec6f4211c4e3c9649,
Sirius allows to define borders with a 0 pix size, it configures the
outline of the Shape figures regarding their border width. 

For regions, we used to set the outline to false once and we add a
specifc one line border to draw the region separator. Now the
RegionRoundedGradientRectangle always disable the ouline of the shape. 

RegionRoundedGradientRectangle is also able to draw two different corner
size to avoid Region to overlap the RegionContainer rounded corners.

Bug: 470770
Change-Id: Ib52248f4711c5ad1d83090f17b611824ff9e035a
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>