[465661] Reset part, policies and figures childrenPresentation change
When the children presentation of a ContainerMapping changes from List
to others ContainerLayout or from others to List, the Sirius model
element are recreate from DNodeList to DNodeContainer or from
DNodeContainer to DNodeList. Their children are also impacted. Then the
corresponding parts see that their View.getElement() has changed and
they call handleMajorSemanticChange to ask the parent to reset the part,
policies and figures. 

When the changes is done between FreeFrom, HStack or VStack, the
DNodeContainer is not replaced but only sees it childrenPresentation
updated. Then a lot of changes are needed on the part, compartment,
content panes, figures, policys, layout manager so we also call
handleMajorSemanticChange to be sure to forget nothing.

Bug: 465661
Change-Id: I4d579b6bde492556f179cbe90d479b621494eb8f
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>