[470822] Add NON-NLS markers missed by EMF regen
Mark NON-NLS all technical/internal strings in src-gen folders which
were not handled by EMF's generation itself. This includes:
* *Package.{eNAME,eNS_URI,eNS_PREFIX};
* constructor parameters (name and literal) for all EEnum values;
* most default values (but not all);
* strings in @not-generated methods (ignored by EMF).

Some string literals in these src-gen folders are not marked NON-NLS,
because they will need to be properly externalized (this will be done as
part of #472057).

Note that now that they have been added, these markers are kept when
re-generating again, so this is a one-shot operation.

Bug: 470822
Change-Id: Ib7958bc1d6e845d901e60f10f992fcc22109a020
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>