[474807] Reinit parsed expressions cache when a new metamodel is known
The parsedExpressions cache is used to avoid re-parsing for queries
which have the same String content, but if a new metamodel is registered 
we should drop this cache as the parsing result could be different. 

As parsing is very fast anyway and activateMetamodels() is not supposed
to be called in "query intensive" times like the representation refresh
for instance, no specific impact analysis is done and all the AST's are
dropped from the cache.

Bug: 474807
Change-Id: Iff15c7bcd41137f97956e5fbb157d9c8a72d3047
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>