[465328] Bracket edge cases-Moving segment
This commit only deals with the moving of the segment of bracket edges.
Other cases will be handled in another commit:
* moving the source or target node of bracket edge
* change orientation of the bracket edge

The EdgeLabelsComputationUtil constructor has now a new parameter to
determine if the label is a center label on a BracketEdge. Indeed, this
label has a specific reference point (see
BracketLabelLocator.getReferencePoint() and
BracketResizableLabelLocator.getReferencePoint()).
The BracketLabelLocator and BracketResizableLabelLocator have been
refactored to reuse similar code. The method getReferencePoint is now
also used in EdgeLabelsComputationUtil when necessary.

The EdgeLabelsComputationUtil has also been completed to handle
correctly the specific case of edge with 2 "inverse" segments on same
line (getVectorFromOldToNewForSegmentsOnSameLine method).

The ResizableShapeLabelEditPolicy now uses EdgeLabelsComputationUtil
instead of LabelHelper for label on connection.

Bug: 465328
Change-Id: Ic8e6ba1aa53f7c3e931df96521b2cdcfbf19cc33
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>