[test] Avoid NPE occuring during DNodeList font color refresh in undo()
This NPE seems to occur only during in 
KeyboardDeletionFromLabelTests.testDeleteFromLabelWithStandardTool(), 
I did not succeed to reproduce with manual steps.

Remove dead code: 
 . DNodeListElementEditPart will never have children
parts
 . Add isActive() guard before calling refresh(). This should be similar
to a getParent() call. This refresh call seems weird: the refresh() is
already called at the begining of the handleNotificationEvent() method.
Fursthermore, ADD_MANY and REMOVE_MANY cases are forgotten.

The NPE occured because an inactive part receive notification from its
GMF node, but it should have been removed from the DiagramEventBroker
listener during deactivate().

I did not find why this is not the case when the test fails.

Change-Id: I4b40f842df0be370fe2a32b4411bf6ff0cae3da0
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>