[470860] Enable collapse for Regions
Enable collapse of Region in both stack directions except when the
RegionContainer is itself a Region.

Store (restore) the expanded bounds in (from) an AbsoluteBoundsFilter.

Vertical Stacks: update the GMF height to -1 to let the figure adapt
itself to the text height

Horizontal Stacks: update the GMF width to -1 when the label is not
truncated to let the figure adapt its width to the label size or current
wrapped state.

Update the tooltip when the region is collapsed with it label hidden:
"label - tooltip".

Forbid resize for collapsed regions: when they are direcly resized or
indirectly resized by their parent region container or sibling regions.

Bug: 470860
Change-Id: I69c3a35e3bcf1bc8ce32babaa953bb5944288a1c
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>