[460610] Ignore selection changed event out of active part
* For instance, the user selects in model explorer view several elements
with a sub-set of them in the representation but an other part which is
not. In that case, elements present in the representation were selected
and the selection change listener was notified of a selection change in
the representation. That caused to select (in the model explorer view)
only items present in the diagram (which is not expected).
* This commit deactivate the associated tests since they are unreliable.

Bug: 460610
Change-Id: I07f13f7e117643710bec132604f5e9bd9f91ab34
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>