[442268] Move utility method into internal helper class
This is technically an API change, as it removed a protected method from
a non-final public class, but the code changes so much more in the next
patches of this series that it is not worth documenting each and every
intermediate API change. Only the final, aggegrate changes will be
documented.

Bug: 442268
Change-Id: I2ff793edfed14d7f89d0a786ca7c1521880a1d79
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>