[460075] Avoid NPE during tooltip refresh
A potential NPE has been detected without reproduction use case. See [1]
for the corresponding stack.
This commit also fixes the test TooltipProviderTests: locally I have a
ClassCastException for testTooltipOnTableEditionDialect (the
allRepresentations.get(1) is a DTree).
The test testTooltipOnModelExplorer() has been temporary removed as it
fails on gerrit but not locally (the view IModelExplorerView.ID is not
found).
The condition for this test, in suite, has also been changed. Indeed,
this test must only be launched if there is currently no
IToolTipProvider already registered (the test provides its own
IToolTipProvider).

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=460075#c6

Bug: 460075
Change-Id: I038d204b510fe446ccf0f4c9f102f0342be8b724
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>