[474878] Make ODesignGenericInterpreter implement IInterpreterWithDiagnostic
ODesignGenericInterpreter is, among all the IInterpreter
implementations, the actual façade visible from outside a Sirius
session. Make it implement IInterpreterWithDiagnostic so that the
additional features this interface provides (better error reporting) is
actually available to every client code which is aware of this feature.
In particular, this makes the finer error reporting available from the
"Interpreter" view.

Bug: 474878
Change-Id: I270bda79f2fe3a3f6d89dcb509a8e64bd2e23ada
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>