[cleanup] Make sure the UI queue is empty after IEditingSession.close()
Calling IEditingSession.close*() forces the closing of any editor related
to the session. But in the context of a JUnitTest the SWT Thread will
not process the corresponding UI events until it is processing the event
queue. If this gets defered, then the editor (which should have been
closed long ago) gets re-activated and might trigger the re-opening of
the Session (which got closed before)

This commit make sure in every place we close the IEditingSession we
also call TestsUtil.emptyEventsFromUIThread(); to make sure the closing
happens right after the call.

Change-Id: I310a256038441a4c9accfb300b97e9e9e11ec38b
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>
Fast-tracked-by: Pierre-Charles David <pierre-charles.david@obeo.fr>