[479019] Introduce and use a MetamodelDescriptorProvider2 API
This API extends the original MetamodelDescriptorProvider API so that 
it can be called even when no Viewpoint is enabled. This allow
implementers to declare accessible metamodels in a systematic way and is
used by Sirius itself so that the Ecore models capturing representation
data are known by the interpreters even if no Viewpoint has been
selected.

Bug: 479019
Change-Id: I6243f43470522c3e1f6496bde4ac63c9a890e0ff
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>