[470859] Complete release note and documentation
Remove the capability to cancel resize propagation on
RegionContainer resize as this is correctly managed from Region
and this produced unwanted side effects for Region which are also
RegionContainers.

Add a Warning validation rule to detect reconnect tools on edge mapping
related to Regions as the reconnect is disabled On Regions.

Bug: 470859
Change-Id: I5a879a451164cfd6cbbba9d3ce8dc542438b0720
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>