[cleanup] Update generated property views formatting
Update formatting/ordering of some of the generated property views code
to match what is actually generated. This does not change anything
substantial to the code, but make the re-generation a no-op for almost
all files except the plugin.xml (which contain the many manual
customizations for ordering and categorization).

Change-Id: I3d014b91bb40ca664827f4e305661f5b3be3700c
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>