[449357] Reincorporate code removed by error
Since the inheritance changes in 3e7f5ad187ff0e6f42, we lost some custom
code from GMF which silenced specific exceptions during resource
loading. As a result, these can now be visible to client code which does
not expect them and is not ready to handle them.

Restore the old treatment of these exceptions as before (it does not
require an actual dependency to GMF code).

For reference, the exceptions which are silenced,
UnresolvedReferenceException, correspond to situations that GMF (and
Sirius) can correctly handle by themselves (by being proxy-aware) and
thus should not be reported as loading errors to clients.

Bug: 449357
Change-Id: I0cce8cb952f77c8ae9a9b936db8e82c14c3d0282
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>