[470867] Test reorder from semantic changes and session status
Remove the specified region size in the test VSM and update the
initial layout test.

Fix the first update bounds operation during editor initialization:
there is no edit part yet, looking for the location of the next region
allows to retrieve the correct previous bounds if they exists
instead of using the default bounds. This avoids to make the session
dirty when opening a diagram editor with auto-sized regions. 

Bug: 470867
Change-Id: I4d9b4674b733e17de83ff1b11ba6b00cb9ac814c
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>