[481299] Correctly manage changes between SVG and others image kinds
If the image path is changed from a SVG to another format (or inverse),
a new figure is now created instead of refresh the existing one.
The test SetStyleToWorkspaceImageTests has been completed to test the
fix. The data of this test has been completed with some nodes with
workspace image style.

Bug: 481299
Change-Id: I634473b75fe555e3005aa16d866316767e0639fb
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>