[481733] Fix CCE and make the code more robust
We customize the *ItemProvider#getLabel() methods of all types which
inherit from IdentifiedElement, but at least in commit
2bff4d82eecc048b5f8fa4d519c71361858e6de5 these customizations are buggy
and downcast the input object to the wrong type (see in particular in
ValidationRuleItemProvider).

Fix the broken casts, and more generally always cast to the more general
IdentifiedElement type (which is all that we need) instead of each
provider casting to the more precise specific type it handles. This
makes the code more similar across all the custom getLabel() and less
brittle.

Bug: 481733
Change-Id: Ib03b5ce7c2ddc33ed8b4425411a0e45793ac5503
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>