[463227] Convert remaining A2 expressions into AQL
With these translations, the test
org.eclipse.sirius.tests.swtbot.sequence.InteractionUseCoverageTests.testCoveredParticipantsChangeByRemovingAExistingParticipant2()
is KO. The translated A2 expressions were false since the beginning (no
"%" the "<"). The test has been written on this wrong base, so it must
be fixed by changing the expected bounds.
The fix reveals another problem in the VSM: "Impossible to find an
interpreter" caused by successive if on removed element.

Bug: 463227
Change-Id: I0c84c072d8c767904a0b53c4a1107830c8cd0120
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>