[test] Review logic of one test of BorderMarginTest
The test
BorderMarginTest.testFixedSizeFromAutoSizeDoNotDisplayScrollBars() is
unreliable according to OS on which it is launched. The logic of this
test has been reviewed. The size is now set according to autosize
diagram instead of using predefined sizes.
Also fix the testAutoSize() test for Mac, that have label smaller that
the minimum 150 pixels of width for container (40 pixels for list).

Change-Id: Ief75d4d55f8b5078cf15f888fce6b112f1a527da
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>